Sync local copy

+++ src/jobs/postjobaction.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2020 Meltytech, LLC
+ * Copyright (c) 2018-2021 Meltytech, LLC
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -19,6 +19,7 @@
 #include "mainwindow.h"
 #include "docks/playlistdock.h"
 #include "shotcut_mlt_properties.h"
+#include <Logger.h>
 
 // For file time functions in FilePropertiesPostJobAction::doAction();
 #include <utime.h>
@@ -46,40 +47,70 @@ void FilePropertiesPostJobAction::doAction()
 #endif
 }
 
-void ReverseOpenPostJobAction::doAction()
+void OpenPostJobAction::doAction()
 {
     FilePropertiesPostJobAction::doAction();
-    QFile::remove(m_fileNameToRemove);
+    if (!m_fileNameToRemove.isEmpty()) {
+        QFile::remove(m_fileNameToRemove);
+    }
     MAIN.open(m_dstFile);
     MAIN.playlistDock()->on_actionAppendCut_triggered();
 }
 
-void ReverseReplacePostJobAction::doAction()
+void ReplaceOnePostJobAction::doAction()
 {
     FilePropertiesPostJobAction::doAction();
-    QFile::remove(m_fileNameToRemove);
-    Mlt::Producer producer(MLT.profile(), m_dstFile.toUtf8().constData());
-    if (producer.is_valid()) {
-        if (!qstrcmp(producer.get("mlt_service"), "avformat")) {
-            producer.set("mlt_service", "avformat-novalidate");
-            producer.set("mute_on_pause", 0);
-        }
-        MLT.lockCreationTime(&producer);
-        producer.set_in_and_out(m_in, -1);
-        MAIN.replaceInTimeline(m_uuid, producer);
+    if (!m_fileNameToRemove.isEmpty()) {
+        QFile::remove(m_fileNameToRemove);
+    }
+    Mlt::Producer newProducer(MLT.profile(), m_dstFile.toUtf8().constData());
+    if (newProducer.is_valid()) {
+        Mlt::Producer* producer = MLT.setupNewProducer(&newProducer);
+        producer->set_in_and_out(m_in, -1);
+        MAIN.replaceInTimeline(m_uuid, *producer);
+        delete producer;
     }
 }
 
-void ConvertReplacePostJobAction::doAction()
+void ReplaceAllPostJobAction::doAction()
 {
     FilePropertiesPostJobAction::doAction();
-    Mlt::Producer producer(MLT.profile(), m_dstFile.toUtf8().constData());
-    if (producer.is_valid()) {
-        if (!qstrcmp(producer.get("mlt_service"), "avformat")) {
-            producer.set("mlt_service", "avformat-novalidate");
-            producer.set("mute_on_pause", 0);
+    Mlt::Producer newProducer(MLT.profile(), m_dstFile.toUtf8().constData());
+    if (newProducer.is_valid()) {
+        Mlt::Producer* producer = MLT.setupNewProducer(&newProducer);
+        MAIN.replaceAllByHash(m_hash, *producer);
+        delete producer;
+    }
+}
+
+void ProxyReplacePostJobAction::doAction()
+{
+    QFileInfo info(m_dstFile);
+    QString newFileName = info.path() + "/" + info.baseName() + "." + info.suffix();
+    if (QFile::rename(m_dstFile, newFileName)) {
+        Mlt::Producer newProducer(MLT.profile(), newFileName.toUtf8().constData());
+        if (newProducer.is_valid()) {
+            Mlt::Producer* producer = MLT.setupNewProducer(&newProducer);
+            producer->set(kIsProxyProperty, 1);
+            producer->set(kOriginalResourceProperty, m_srcFile.toUtf8().constData());
+            MAIN.replaceAllByHash(m_hash, *producer, true);
+            delete producer;
+        } else {
+            LOG_WARNING() << "proxy file is invalid" << newFileName;
+            QFile::remove(m_dstFile);
         }
-        MLT.lockCreationTime(&producer);
-        MAIN.replaceAllByHash(m_hash, producer);
+    } else {
+        LOG_WARNING() << "failed to rename" << m_dstFile << "as" << newFileName;
+        QFile::remove(m_dstFile);
+    }
+}
+
+void ProxyFinalizePostJobAction::doAction()
+{
+    QFileInfo info(m_dstFile);
+    QString newFileName = info.path() + "/" + info.baseName() + "." + info.suffix();
+    if (!QFile::rename(m_dstFile, newFileName)) {
+        LOG_WARNING() << "failed to rename" << m_dstFile << "as" << newFileName;
+        QFile::remove(m_dstFile);
     }
 }
